Ensure entire SIP body is read
RFC 3261 states that the size of the message body does not include the
CRLF separating header fields and the body. However, the CRLF was not
being skipped, and instead was being included in the payload of the
RawMessage. This had the effect of cutting off the last two characters
of the message body.